Fixed Response.setData. Work on example (comments, categories). Introduced widgets concept (Aside) implemented as standard apps (better name needed?) Other fixes.

+++ example/src/root/id.js
@@ -1,15 +1,17 @@
-var Article = require("blog/article").Article;
-
+var Article = require("blog/article").Article,
+    Comment = require("blog/comment").Comment;
+    
 exports.app = function(request, response) {
     var params = request.params();
 
-    switch (request.requestMethod()) {
-        case "GET":
-            response.setData({
-                article: $db.query("SELECT * FROM Article WHERE id=?", params.id).one(Article)
-            });
-        case "DELETE":
-            $db.execute("DELETE FROM Article WHERE id=?", params.id);
-            response.redirect();
+    if (request.isGet()) {
+        response.setData({
+            article: $db.query("SELECT * FROM Article WHERE id=?", params.id).one(Article),
+            comments: $db.query("SELECT * FROM Comment WHERE parentId=?", params.id).all(Comment)
+        });
+    } else if (request.isDelete()) {
+        $db.execute("DELETE FROM Article WHERE id=?", params.id);
+        response.redirect();
+    } else /* isPut() */ {
     }
 }
